+++ src/main/java/de/maxhenkel/easyvillagers/ClientConfig.java
@@ -7,6 +7,7 @@ public class ClientConfig extends ConfigBase {
 
     public final ForgeConfigSpec.BooleanValue enableRightClickPickup;
     public final ForgeConfigSpec.DoubleValue villagerVolume;
+    public final ForgeConfigSpec.EnumValue<CycleTradesButtonLocation> cycleTradesButtonLocation;
 
     public ClientConfig(ForgeConfigSpec.Builder builder) {
         super(builder);
@@ -17,6 +18,14 @@ public class ClientConfig extends ConfigBase {
         villagerVolume = builder
                 .comment("The volume of every villager related sound in this mod")
                 .defineInRange("villager.volume", 1D, 0D, 1D);
+
+        cycleTradesButtonLocation = builder
+                .comment("The location of the cycle trades button")
+                .defineEnum("villager.cycle_trades_button_location", CycleTradesButtonLocation.TOP_LEFT);
+    }
+
+    public enum CycleTradesButtonLocation {
+        TOP_LEFT, TOP_RIGHT
     }
 
 }